Fix inverted dependency.
