+++ torch-sys/libtch/torch_api.cpp
@@ -288,14 +288,6 @@ void at_copy_(tensor dst, tensor src) {
   )
 }
 
-extern "C" bool tch_write_stream_destructor(void *stream_ptr);
-extern "C" bool tch_write_stream_write(void *stream_ptr, const uint8_t *buf, size_t size, size_t *out_size);
-extern "C" bool tch_read_stream_destructor(void *stream_ptr);
-extern "C" bool tch_read_stream_stream_position(void *stream_ptr, uint64_t *pos);
-extern "C" bool tch_read_stream_seek_start(void *stream_ptr, uint64_t pos, uint64_t *new_pos);
-extern "C" bool tch_read_stream_seek_end(void *stream_ptr, int64_t pos, uint64_t *new_pos);
-extern "C" bool tch_read_stream_read(void *stream_ptr, uint8_t *buf, size_t size, size_t *new_pos);
-
 namespace {
   class WriteStreamAdapter {
   private:
