+++ torch-sys/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod io;
+
 use libc::{c_char, c_int, c_uchar, c_void, size_t};
 
 #[repr(C)]
